Vector3Dを使うように修正
その他変更

+++ Cont/Shootingood/Enemy.cpp
@@ -37,47 +37,47 @@ Enemy::~Enemy() {}
 void Enemy::Render() {	
 	// 自分を描画
 	if (!enemy_status->Is_Damage()) {
-		DrawExtendGraph(vector3->x, vector3->y, vector3->x + width, 
-						vector3->y + height, enemy_graph, FALSE);
+		DrawExtendGraph((int)vector3->x, (int)vector3->y, (int)vector3->x + width,
+			(int)vector3->y + height, enemy_graph, FALSE);
 		return;
 	}
 	if (enemy_status->Is_Damage()) {
-		DrawExtendGraph(vector3->x, vector3->y, vector3->x + width, 
-						vector3->y + height, damage_enemy_graph, FALSE);
+		DrawExtendGraph((int)vector3->x, (int)vector3->y, (int)vector3->x + width,
+			(int)vector3->y + height, damage_enemy_graph, FALSE);
 	}
 }
 
 // エネミーの右端の座標を返す
-int Enemy::Get_Right_Edge() {
+float Enemy::Get_Right_Edge() {
 	return vector3->x + width;
 }
 
 // エネミーの左端の座標を返す
-int Enemy::Get_Left_Edge() {
+float Enemy::Get_Left_Edge() {
 	return vector3->x;
 }
 
 // エネミーの上端の座標を返す
-int Enemy::Get_Top_Edge() {
+float Enemy::Get_Top_Edge() {
 	return vector3->y;
 }
 
 // エネミーの下端の座標を返す
-int Enemy::Get_Bottom_Edge() {
+float Enemy::Get_Bottom_Edge() {
 	return vector3->y + height;
 }
 
 // 弾の発射する位置を返す
-int Enemy::Get_Shot_Point() {
+float Enemy::Get_Shot_Point() {
 	return vector3->x + (width / 2);
 }
 
 // 初期x座標を返す
-int Enemy::Get_Initialize_Position_X() {
+float Enemy::Get_Initialize_Position_X() {
 	return initialize_position_x;
 }
 
 // 初期y座標を返す
-int Enemy::Get_Initialize_Position_Y() {
+float Enemy::Get_Initialize_Position_Y() {
 	return initialize_position_y;
 }
