+++ Cont/Shootingood/Player.h
@@ -21,16 +21,16 @@ public:
 	void Render() override;
 
 	// アクターの右端の座標を返す
-	virtual int Get_Right_Edge() override;
+	virtual float Get_Right_Edge() override;
 	// アクターの左端の座標を返す
-	virtual int Get_Left_Edge() override;
+	virtual float Get_Left_Edge() override;
 	// アクターの上端の座標を返す
-	virtual int Get_Top_Edge() override;
+	virtual float Get_Top_Edge() override;
 	// アクターの下端の座標を返す
-	virtual int Get_Bottom_Edge() override;
+	virtual float Get_Bottom_Edge() override;
 
 	// 弾が出る位置を返す
-	int Get_Shoot_Point();
+	float Get_Shoot_Point();
 
 	// パワーアップ状態であるかをセットする
 	void Set_Power_Up(bool set_is_power_up);
